GH-1277: SafeNotifyService threads leak in CuratorFrameWorkImpl (#1278)
CURATOR-495 introduced a new runSafeService field in CuratorFrameworkImpl class, and this field is
either initialized by an external ExecutorService via the builder, or it is created internally
within the class.

In the CuratorFrameworkImpl#close method though, this Executor is never closed, so the threads
that are opened by the instances are lingering there until the VM is closed by default.
Worse, if someone specifies a thread factory to the framework implementation via the builder that
produces non-daemon threads, the VM never exits due to the unstopped single thread executor.

Fixes #1277.

(cherry picked from commit f18f0acf432231ece5dcd179020c4951dbe49803)